Extract 360dialog to its own controller
