+++ app/controllers/whats_app/webhook_controller.rb
@@ -1,9 +1,9 @@
 # frozen_string_literal: true
 
-# rubocop:disable Metrics/ClassLength, Metrics/AbcSize
-# TODO: Separate WhatsApp implementations intro separate controllers
 module WhatsApp
   class WebhookController < ApplicationController
+    include HandleCallbacks
+
     skip_before_action :require_login, :verify_authenticity_token
     UNSUCCESSFUL_DELIVERY = %w[undelivered failed].freeze
 
@@ -38,41 +38,6 @@ module WhatsApp
       adapter.consume(whats_app_message_params) { |message| message.contributor.reply(adapter) }
     end
 
-    def three_sixty_dialog_message
-      head :ok
-      return if params['statuses'].present? # TODO: Do we want to handle statuses?
-
-      handle_error(params['errors']) if params['errors'].present?
-
-      adapter = WhatsAppAdapter::ThreeSixtyDialogInbound.new
-
-      adapter.on(WhatsAppAdapter::UNKNOWN_CONTRIBUTOR) do |whats_app_phone_number|
-        handle_unknown_contributor(whats_app_phone_number)
-      end
-
-      adapter.on(WhatsAppAdapter::REQUEST_FOR_MORE_INFO) do |contributor|
-        handle_request_for_more_info(contributor)
-      end
-
-      adapter.on(WhatsAppAdapter::REQUEST_TO_RECEIVE_MESSAGE) do |contributor|
-        handle_request_to_receive_message_three_sixty_dialog(contributor)
-      end
-
-      adapter.on(WhatsAppAdapter::UNSUPPORTED_CONTENT) do |contributor|
-        WhatsAppAdapter::Outbound.send_unsupported_content_message!(contributor)
-      end
-
-      adapter.on(WhatsAppAdapter::UNSUBSCRIBE_CONTRIBUTOR) do |contributor|
-        handle_unsubsribe_contributor(contributor)
-      end
-
-      adapter.on(WhatsAppAdapter::SUBSCRIBE_CONTRIBUTOR) do |contributor|
-        handle_subscribe_contributor(contributor)
-      end
-
-      adapter.consume(three_sixty_dialog_message_params.to_h) { |message| message.contributor.reply(adapter) }
-    end
-
     def errors
       return unless error_params['Level'] == 'ERROR'
 
@@ -105,27 +70,6 @@ module WhatsApp
                     :ReferralNumMedia, :SmsMessageSid, :SmsSid, :SmsStatus, :To, :WaId)
     end
 
-    def three_sixty_dialog_message_params
-      params.permit({ webhook: [contacts: [:wa_id, { profile: [:name] }],
-                                messages: [:from, :id, :type, :timestamp, { text: [:body] }, { context: %i[from id] },
-                                           { button: [:text] }, { image: %i[id mime_type sha256] }, { voice: %i[id mime_type sha256] },
-                                           { video: %i[id mime_type sha256] },
-                                           { errors: %i[code details title] },
-                                           { location: %i[latitude longitude timestamp type] },
-                                           { contacts: [{ org: {} }, { addresses: [] }, { emails: [] }, { ims: [] },
-                                                        { phones: %i[phone type wa_id] }, { urls: [] },
-                                                        { name: %i[first_name formatted_name last_name] }] }]] },
-                    contacts: [:wa_id, { profile: [:name] }],
-                    messages: [:from, :id, :type, :timestamp, { text: [:body] }, { context: %i[from id] }, { button: [:text] },
-                               { image: %i[id mime_type sha256] }, { voice: %i[id mime_type sha256] },
-                               { video: %i[id mime_type sha256] },
-                               { errors: %i[code details title] },
-                               { location: %i[latitude longitude timestamp type] },
-                               { contacts: [{ org: {} }, { addresses: [] }, { emails: [] }, { ims: [] },
-                                            { phones: %i[phone type wa_id] }, { urls: [] },
-                                            { name: %i[first_name formatted_name last_name] }] }])
-    end
-
     def error_params
       params.permit(:AccountSid, :Level, :ParentAccountSid, :Payload, :PayloadType, :Sid, :Timestamp)
     end
@@ -151,12 +95,6 @@ module WhatsApp
 
       message = (send_requested_message(contributor, twilio_message_sid) if twilio_message_sid)
       WhatsAppAdapter::Outbound.send!(message || contributor.received_messages.first)
-    end
-
-    def handle_request_to_receive_message_three_sixty_dialog(contributor)
-      contributor.update!(whats_app_message_template_responded_at: Time.current, whats_app_message_template_sent_at: nil)
-
-      WhatsAppAdapter::Outbound.send!(contributor.received_messages.first)
     end
 
     def handle_unsubsribe_contributor(contributor)
@@ -196,11 +134,5 @@ module WhatsApp
       ErrorNotifier.report(e)
       nil
     end
-
-    def handle_error(error)
-      exception = WhatsAppAdapter::ThreeSixtyDialogError.new(error_code: error['code'], message: error['title'])
-      ErrorNotifier.new(exception, context: { details: error['details'] })
-    end
   end
 end
-# rubocop:enable Metrics/ClassLength, Metrics/AbcSize
