+++ config/routes.rb
@@ -48,9 +48,8 @@ Rails.application.routes.draw do
     post '/webhook', to: 'webhook#message'
     post '/errors', to: 'webhook#errors'
     post '/status', to: 'webhook#status'
-    post '/three-sixty-dialog-status', to: 'webhook#three_sixty_dialog_status'
     get '/onboarding-successful', to: 'onboarding#success'
-    post '/three-sixty-dialog-webhook', to: 'webhook#three_sixty_dialog_message'
+    post '/three-sixty-dialog-webhook', to: 'three_sixty_dialog_webhook#message'
   end
 
   telegram_webhook Telegram::WebhookController
